feat: show chat UI on home page
