+++ src/app/page.tsx
@@ -1,9 +1,6 @@
+import Chat from "@/components/Chat";
+
 export default function Home() {
-  return (
-    <div className="flex items-center justify-center min-h-screen bg-background text-foreground">
-      <div className="text-center">
-        <h1 className="text-4xl md:text-6xl font-bold">Nature&apos;s agent is coming soon</h1>
-      </div>
-    </div>
-  );
+  return <Chat />;
 }
+
